Fixed checkOscillation function
separated out stripGraphics for use in both newCycle and checkOscillation (checkOscillation applies it to a deepcopy of board so as not to affect the main board).

Also changed things so that graphics set can be changed in settings. Currently using an H-based display as it gives a better sense of motion than the X-based set.

Separated out settings, initialisations and main loop to make it more readable.