Graphics run through a printBoard function
Rather than strip the graphics in order to then update the board, the entire board is now represented only using 'X' or ' ' and then the printBoard function uses the history array (previously just used to detect oscillators) to decide whether a cell is newborn or dying. 

Dumbest thing now is that I didn't think to just have "True" and "False" as the values stored in the board, but I guess it doesn't make a huge difference.